1.4.0 DEMO
This is the demo for 1.4.0 This will add parts and "will add" Later.
* Partial spanish recognition

* More responses

* Some fixs

* Can now ask Siri for data values and crafting recipes (Partial)

And more. 

Just to let you know I am working hard and can hardly add all the features as fast as the other updates as schoolwork is topping it.

Since this update It has add over 400 lines of code. So wait. ;)